+++ src/main/java/kitchenpos/menus/tobe/domain/menu/MenuPrice.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
 
 import jakarta.persistence.Embeddable;
 import java.math.BigDecimal;
