+++ src/main/java/kitchenpos/menus/tobe/domain/menugroup/MenuGroupId.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menugroup;
 
 import jakarta.persistence.Embeddable;
 import java.util.Objects;
