+++ src/main/java/kitchenpos/menus/tobe/domain/menuproduct/MenuProduct.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menuproduct;
 
 import jakarta.persistence.AttributeOverride;
 import jakarta.persistence.Column;
