+++ src/main/java/kitchenpos/menus/tobe/domain/menuproduct/MenuProductQuantity.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menuproduct;
 
 import jakarta.persistence.Embeddable;
 import java.util.Objects;
