+++ src/main/java/kitchenpos/menus/tobe/infra/persistence/menu/TobeJpaMenuRepository.java
@@ -1,9 +1,9 @@
-package kitchenpos.menus.tobe.infra.persistence;
+package kitchenpos.menus.tobe.infra.persistence.menu;
 
 import java.util.List;
 import java.util.UUID;
-import kitchenpos.menus.tobe.domain.Menu;
-import kitchenpos.menus.tobe.domain.TobeMenuRepository;
+import kitchenpos.menus.tobe.domain.menu.Menu;
+import kitchenpos.menus.tobe.domain.menu.TobeMenuRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
