+++ src/main/java/kitchenpos/menus/tobe/infra/persistence/menugroup/TobeJpaMenuGroupRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.menus.tobe.infra.persistence;
+package kitchenpos.menus.tobe.infra.persistence.menugroup;
 
 import java.util.UUID;
-import kitchenpos.menus.tobe.domain.MenuGroup;
-import kitchenpos.menus.tobe.domain.TobeMenuGroupRepository;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroup;
+import kitchenpos.menus.tobe.domain.menugroup.TobeMenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TobeJpaMenuGroupRepository extends TobeMenuGroupRepository,
