+++ src/test/java/kitchenpos/menus/tobe/domain/MenuGroupIdTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 import java.util.UUID;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroupId;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
