+++ src/test/java/kitchenpos/menus/tobe/domain/MenuGroupNameTest.java
@@ -3,6 +3,7 @@ package kitchenpos.menus.tobe.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroupName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
