+++ src/test/java/kitchenpos/menus/tobe/domain/MenuGroupTest.java
@@ -3,6 +3,7 @@ package kitchenpos.menus.tobe.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.UUID;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
