+++ src/test/java/kitchenpos/menus/tobe/domain/MenuNameTest.java
@@ -3,6 +3,7 @@ package kitchenpos.menus.tobe.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
+import kitchenpos.menus.tobe.domain.menu.MenuName;
 import kitchenpos.mock.client.FakePurgomalumClient;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
