+++ src/test/java/kitchenpos/menus/tobe/domain/MenuProductTest.java
@@ -3,6 +3,8 @@ package kitchenpos.menus.tobe.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.math.BigDecimal;
+import kitchenpos.menus.tobe.domain.menuproduct.MenuProduct;
+import kitchenpos.menus.tobe.domain.menuproduct.MenuProductQuantity;
 import kitchenpos.mock.fixture.ProductFixture;
 import kitchenpos.products.tobe.domain.Product;
 import org.junit.jupiter.api.DisplayName;
