+++ src/test/java/kitchenpos/menus/tobe/domain/MenuTest.java
@@ -3,6 +3,7 @@ package kitchenpos.menus.tobe.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.math.BigDecimal;
+import kitchenpos.menus.tobe.domain.menu.Menu;
 import kitchenpos.mock.fixture.MenuFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
